+++ csvkit/utilities/csvclean.py
@@ -14,6 +14,15 @@ class CSVClean(CSVKitUtility):
 
     def add_arguments(self):
         self.argparser.add_argument(
+            '--length-mismatch', dest='length_mismatch', action='store_true',
+            help='Report data rows that are shorter or longer than the header row.')
+        self.argparser.add_argument(
+            '--empty-columns', dest='empty_columns', action='store_true',
+            help='Report empty columns as errors.')
+        self.argparser.add_argument(
+            '-a', '--enable-all-checks', dest='enable_all_checks', action='store_true',
+            help='Enable all error reporting.')
+        self.argparser.add_argument(
             '--header-normalize-space', dest='header_normalize_space', action='store_true',
             help='Strip leading and trailing whitespace and replace sequences of whitespace characters by a single '
                  'space in the header.')
@@ -29,27 +38,42 @@ class CSVClean(CSVKitUtility):
         self.argparser.add_argument(
             '--fillvalue', dest='fillvalue',
             help='The value with which to fill short rows. Defaults to none.')
-        self.argparser.add_argument(
-            '--empty-columns', dest='empty_columns', action='store_true',
-            help='Report empty columns as errors.')
 
     def main(self):
         if self.additional_input_expected():
             sys.stderr.write('No input file or piped data provided. Waiting for standard input:\n')
 
+        if (
+            # Checks
+            not self.args.length_mismatch
+            and not self.args.empty_columns
+            and not self.args.enable_all_checks
+            # Fixes
+            and not self.args.header_normalize_space
+            and not self.args.join_short_rows
+            and not self.args.fill_short_rows
+        ):
+            self.argparser.error('No checks or fixes were enabled. See available options with: csvclean --help')
+
         if self.args.join_short_rows and self.args.fill_short_rows:
             self.argparser.error('The --join-short-rows and --fill-short-rows options are mutually exclusive.')
+
+        default = self.args.enable_all_checks
 
         reader = agate.csv.reader(self.skip_lines(), **self.reader_kwargs)
 
         checker = RowChecker(
             reader,
+            # Checks
+            length_mismatch=default or self.args.length_mismatch,
+            empty_columns=default or self.args.empty_columns,
+            # Fixes
             header_normalize_space=self.args.header_normalize_space,
             join_short_rows=self.args.join_short_rows,
             separator=self.args.separator,
             fill_short_rows=self.args.fill_short_rows,
             fillvalue=self.args.fillvalue,
-            empty_columns=self.args.empty_columns,
+            # Other
             zero_based=self.args.zero_based,
         )